+++ src/hkimjp/konpy2/comments.clj
@@ -22,7 +22,12 @@
                 :updated (now)})
       (r/after-comment (user request))
       (redirect (str "/k/problem/" pid)))
-    (error-page (user request) "きちんと回答、コメント読んでコメントしないと。")))
+    (page
+     [:div
+      [:div.text-2xl "Error"
+       (when-let [msg (c/get (format "kp2:%s:flash" user))]
+         [:p.text-red-500 msg])
+       [:p  "じゅうぶんに回答・コメント読んでコメントしないと。"]]])))
 
 (defn hx-comment [{{:keys [e]} :path-params}]
   (t/log! {:level :info :id "hx-comment"})